requirements.txt
Django==1.10.6
django-classifier==0.2.1
six
